refactor(deep-thinking): separate CLI and library code
Move server bootstrap code from index.ts to new bin.ts
Update package.json to build both files
Remove side effects from index.ts for better testability
